ref(metrics): change sentry_metrics_count to use int64_t
Per the Sentry metrics specification, counter metrics should use 64-bit
signed integers. This is consistent with sentry_value_new_int64 /
sentry_value_as_int64 in sentry.h.

Also refactored internal record_metric to pass sentry_value_t instead
of double, allowing proper type preservation.

+++ include/sentry.h
@@ -2189,7 +2189,7 @@ typedef enum {
  * request counts or error counts.
  */
 SENTRY_EXPERIMENTAL_API sentry_metrics_result_t sentry_metrics_count(
-    const char *name, double value, const char *unit,
+    const char *name, int64_t value, const char *unit,
     sentry_value_t attributes);
 
 /**
